+++ src/scicloj/noj/v1/vis/python.clj
@@ -10,12 +10,6 @@
                 'matplotlib.backends.backend_agg
                 'numpy)
 
-(defn html-in-reagent [html]
-  (-> [:div
-       {:dangerouslySetInnerHTML
-        {:__html html}}]
-      kind/reagent))
-
 (defmacro with-pyplot
   "Takes forms with mathplotlib.pyplot and returns a showable (SVG) plot."
   [& body]
@@ -28,7 +22,7 @@
      (matplotlib.pyplot/savefig path#)
      (-> path#
          slurp
-         html-in-reagent)))
+         kind/html)))
 
 
 (defn pyplot
@@ -43,4 +37,4 @@
     (matplotlib.pyplot/savefig path)
     (-> path
         slurp
-        html-in-reagent)))
+        kind/html)))