+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [1-alpha35] - unreleased
+- avoding the unnecessary use of Reagent for Python plots
+
 ## [1-alpha34] - 2024-04-15
 - removed the currently-unnecessary `calc-correlations-matrix` and `round` functions (#22) - thanks, @behrica
 - dropped the histogram computation function and used the one in Fastmath instead (#21) - thanks, @behrica